pyproject.toml
[project]
name = "vps-provision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyinfra>=3.4.1",
    "python-dotenv>=1.1.1",
    "testinfra>=6.0.0",
    "pytest>=7.0.0",
]
